documentation added for new expandable recycler adapter classes Added default implementation of parent and child view types removed SimpleExpandableAdapter (it was a dumb idea)
